refactor: update rollout resource name to be nested under plan
This change updates the Rollout resource name to be a singleton sub-resource of Plan (projects/{project}/plans/{plan}/rollout).

- Proto:
  - Updated RolloutService RPCs to use nested paths.
  - Removed redundant 'rollout' field from CreateRolloutRequest.
  - Updated Plan message to use 'has_rollout' boolean.
- Backend:
  - Updated resource name parsing logic in common/resource_name.go.
  - Updated API implementations to use int64 Plan IDs and new resource names.
  - Updated migration script to backfill new resource name format.
  - Fixed tests to align with new API.
- Frontend:
  - Updated CreateRolloutRequest usage.
  - Updated resource name extraction and construction logic.
  - Fixed 404 errors in rollout navigation and detail views.

+++ backend/api/v1/revision_service.go
@@ -205,7 +205,7 @@ func (s *RevisionService) createRevisions(
 				return nil, connect.NewError(connect.CodeNotFound, errors.Errorf("taskRun %q not found", revision.TaskRun))
 			}
 			if taskRun.ProjectID != projectID ||
-				taskRun.PlanUID != int64(rolloutID) ||
+				taskRun.PlanUID != rolloutID ||
 				taskRun.Environment != formatEnvironmentFromStageID(stageID) ||
 				taskRun.TaskUID != taskID {
 				return nil, connect.NewError(connect.CodeNotFound, errors.Errorf("taskRun %q not found", revision.TaskRun))
